refactor: Move pickWebsocketsTemplatePart.js closer to websockets
