+++ lib/plugins/aws/package/compile/events/websockets/lib/deployment.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const crypto = require('crypto');
 const BbPromise = require('bluebird');
-const pickWebsocketsTemplatePart = require('../../../../../utils/pickWebsocketsTemplatePart');
+const pickWebsocketsTemplatePart = require('./pickWebsocketsTemplatePart');
 
 module.exports = {
   compileDeployment() {
