+++ lib/plugins/aws/package/compile/events/websockets/lib/pickWebsocketsTemplatePart.js
